tests: Merge `idl` and `new-idl` tests
