+++ tests/idl/Anchor.toml
@@ -9,7 +9,6 @@ idl = "id11111111111111111111111111111111111111111"
 relations_derivation = "Re1ationsDerivation111111111111111111111111"
 non_existent = { address = "NonExistent11111111111111111111111111111111", idl = "non-existent.json" }
 numbers_123 = { address = "Numbers111111111111111111111111111111111111", idl = "idls/relations.json" }
-new_idl = "Newid11111111111111111111111111111111111111"
 
 [provider]
 cluster = "localnet"
