+++ tests/idl/programs/idl/Cargo.toml
@@ -6,18 +6,12 @@ edition = "2021"
 
 [lib]
 crate-type = ["cdylib", "lib"]
-name = "idl"
 
 [features]
-no-entrypoint = []
-no-idl = []
-no-log-ix-name = []
-cpi = ["no-entrypoint"]
 idl-build = ["anchor-lang/idl-build", "anchor-spl/idl-build", "external/idl-build"]
-default = []
 
 [dependencies]
 anchor-lang = { path = "../../../../lang" }
 anchor-spl = { path = "../../../../spl" }
-bytemuck = {version = "1.4.0", features = ["derive", "min_const_generics"] }
+bytemuck = { version = "*", features = ["derive", "min_const_generics"] }
 external = { path = "../external", features = ["no-entrypoint"] }
