--- tests/idl/programs/new-idl/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "new-idl"
-version = "0.1.0"
-description = "Created with Anchor"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-name = "new_idl"
-
-[features]
-no-entrypoint = []
-no-idl = []
-cpi = ["no-entrypoint"]
-default = []
-idl-build = ["anchor-lang/idl-build"]
-
-[dependencies]
-anchor-lang = { path = "../../../../lang" }
-external = { path = "../external", features = ["no-entrypoint"] }
